Refactored transformer
